+++ crates/sky_ecs/src/ecs/mod.rs
@@ -36,6 +36,7 @@ mod query;
 mod resource;
 mod system;
 pub(crate) mod time;
+mod unwind;
 mod world;
 
 pub use access::{